requirements.txt
websocket-client
pprint
TA-lib
python-binance
datetime
telegram_send
numpy
pandas
backtrader
